// src/pages/dashboard/user/types.ts
export interface UserDataType {
    key: string;
    sno: string;
    userName: string;
    email: string;
    companyName: string;
}

export interface UserDetailsProps {
    name: string;
    email: string;
    contactNumber: string;
    gender: string;
    dateOfBirth: string;
    address: string;
    postcode: string;
    city: string;
    country: string;
}